+++ samples/DLCBuilder/RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 - Added a new validation check for a finger changing during a slide.
 - Added a new validation check for position shift into a pull-off.
 - Fixed false positive in validation for anchor position in relation to chord fingering with chords that use the thumb.
+- Fixed some rare false positives for the "anchor not exactly on note" validation.
 - A better error message is shown when loading an "EXT" vocals file fails.
 - Master ID and persistent ID can now be manually edited also for vocals arrangements when "show advanced features" is enabled.
 - Duplicate tone keys will be prevented when changing the tone key in the UI.